refactor: 디렉토리 구조 변경
- hast-util-to-jsx-runtime/lib/components 라이브러리 설치
- ContestPage 파일 구조 변경
- 디렉토리 구조 변경
