+++ frontend/src/components/submissionResult/types.ts
@@ -1,0 +1,8 @@
+export interface SubmitResult {
+  contestId: number;
+  problemId: number;
+  testcaseId: number;
+  resultStatus: number;
+  elapsedTime: number;
+  memoryUsage: number;
+}
