+++ frontend/styled-system/types/parts.d.ts
@@ -1,4 +1,4 @@
-
+/* eslint-disable */
 export interface Part {
   selector: string
 }
